Add command line option to set the number of threads
The long runtime of run_segalign again is due to the fact that the majority of the wall time is spent in lastz. Each lastz process is run concurrently, but only up to nproc processes. When using srun -n1 -c2 to launch the run_segalign script, this limits the number of concurrent lastz processes to 2. On the other hand, we have found (via strace) that the segalign and segalign_repeat_masker executables hang in a futex (FUTEX_WAIT_PRIVATE) call when more than two threads are used. We suspect this is coming from the tbb flow_graph wait_for_all call.

The segalign and segalign_repeat_masker programs are set up to get the number of threads from tbb::task_scheduler_init::default_num_threads(). When launching the executable underneath srun this will result in the number of threads being set to the same value as passed to the --cpus-per-task/-c flag. Here, we’d like to be able to set the number of threads for the segalign executable independently from the number of concurrent lastz processes.

To work around this issue, we add a command-line-option (--num_threads) for segalign and segalign_repeat_masker that specifies the number of threads for intel-tbb and set its default value to 2. users can revert to using default_num_threads() by setting this command line option to -1.